+++ apps/frontend/src/app/page.tsx
@@ -1,4 +1,5 @@
-import { SprintLogin, WorkSelector } from '@/components/hello-world';
+import { SprintLogin } from '@/components/SprintLogin';
+import { WorkSelector } from '@/components/WorkSelector';
 
 export default function Home() {
   return (
