+++ apps/frontend/src/components/footer.tsx
@@ -1,0 +1,7 @@
+export function Footer() {
+  return (
+    <footer className='bg-page-bg-color flex flex-col items-center justify-center'>
+      <p className='text-gray-500 text-sm'>@Kir-Dev pod-oscar</p>
+    </footer>
+  );
+}
